+++ src/autoload.php
@@ -1,10 +1,10 @@
 <?php
 
-function __autoload_google_spreadsheet__ ($name){
+function autoload_google_spreadsheet ($name){
 	$path = __DIR__ . "/" . str_replace("_", "/", $name) . ".php";
 	if(file_exists($path)){
 		require_once($path);
 	}
 }
 
-spl_autoload_register(__autoload_google_spreadsheet__);
+spl_autoload_register('autoload_google_spreadsheet');