+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 
 # Change Log
 
+## 1.0.0
+
+- rewrite all for Sheets API v4
+- `config()` method moved to `Google_Spreadsheet_Sheet`
+- `update()` method usage has been changed  
+- `Google_Spreadsheet_Sheet` constructor no longer call fetch automatically  
+  call manually `$sheet->fetch()` when starting to use
+
 ## 0.1.5
 
 - improve: let Sheet::update() to receive condition as 1st argument